+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Features:
 Fixes:
 - Add ``cython.final`` to leaf classes, ensuring that they are not subclassed.
 - Warn that ``CodecContext.decode()`` is not memory safe in some cases.
+- Fix ``enumerate_input_devices`` and ``enumerate_output_devices`` raising ``AttributeError`` (:issue:`2264`).
 
 v17.0.1
 -------